requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

# azure-functions
# azure-core
# azure-storage-blob
# boto3
# botocore





azure-core==1.30.2
azure-functions==1.20.0
azure-storage-blob==12.21.0
boto3==1.34.146
botocore==1.34.146
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
idna==3.7
isodate==0.6.1
jmespath==1.0.1
pycparser==2.22
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
